requirements.txt
PyMuPDF==1.23.8
pdfplumber==0.10.3
pandas==2.1.4
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
python-dateutil==2.8.2
openpyxl==3.1.2
Flask==3.0.0
Werkzeug==3.0.1
regex>=2022.0.0
gunicorn==21.2.0 